Main corregido, con av

+++ src/stack_init.c
@@ -30,7 +30,7 @@ static void	stack_free(t_stack_node **stack)
 	put_error();
 }
 
-long	ft_atol(const char *str)
+int	ft_atol(const char *str)
 {
 	long	res;
 	int		sign;
